# perf: Optimize readFromBlock for large inputs