Improved readFromBlock function performance when processing large inputs, achieving 10% speedup by:
- Reducing branch instructions in the hot path
- Minimizing field writes to increase register allocation opportunities

Benchmark results:
Before:
```
goos: darwin
goarch: arm64
pkg: github.com/cosnicolaou/pbzip2/internal/bzip2
cpu: Apple M2 Max
BenchmarkDecodeDigits-12    	     336	   3441370 ns/op	  29.06 MB/s	 3613294 B/op	      51 allocs/op
BenchmarkDecodeDigits-12    	     345	   3458985 ns/op	  28.91 MB/s	 3613308 B/op	      51 allocs/op
BenchmarkDecodeDigits-12    	     345	   3443809 ns/op	  29.04 MB/s	 3613292 B/op	      51 allocs/op
BenchmarkDecodeNewton-12    	      85	  13699671 ns/op	  41.40 MB/s	 3631139 B/op	      51 allocs/op
BenchmarkDecodeNewton-12    	      86	  13215102 ns/op	  42.92 MB/s	 3631142 B/op	      51 allocs/op
BenchmarkDecodeNewton-12    	      87	  13259011 ns/op	  42.78 MB/s	 3631133 B/op	      51 allocs/op
BenchmarkDecodeRand-12      	     985	   1206515 ns/op	  13.58 MB/s	 3644458 B/op	      51 allocs/op
BenchmarkDecodeRand-12      	     990	   1244168 ns/op	  13.17 MB/s	 3644466 B/op	      51 allocs/op
BenchmarkDecodeRand-12      	     944	   1215279 ns/op	  13.48 MB/s	 3644461 B/op	      51 allocs/op
BenchmarkWiktionary-12      	       1	184523727375 ns/op	  56.90 MB/s	367217144 B/op	  542483 allocs/op
BenchmarkWiktionary-12      	       1	188378125708 ns/op	  55.73 MB/s	367217160 B/op	  542483 allocs/op
BenchmarkWiktionary-12      	       1	184545993750 ns/op	  56.89 MB/s	367217144 B/op	  542483 allocs/op
```

After:
```
goos: darwin
goarch: arm64
pkg: github.com/cosnicolaou/pbzip2/internal/bzip2
cpu: Apple M2 Max
BenchmarkDecodeDigits-12    	     363	   3222846 ns/op	  31.03 MB/s	 3613296 B/op	      51 allocs/op
BenchmarkDecodeDigits-12    	     366	   3483579 ns/op	  28.71 MB/s	 3613312 B/op	      51 allocs/op
BenchmarkDecodeDigits-12    	     361	   3212324 ns/op	  31.13 MB/s	 3613296 B/op	      51 allocs/op
BenchmarkDecodeNewton-12    	      92	  13106987 ns/op	  43.27 MB/s	 3631140 B/op	      51 allocs/op
BenchmarkDecodeNewton-12    	      94	  12964625 ns/op	  43.75 MB/s	 3631149 B/op	      51 allocs/op
BenchmarkDecodeNewton-12    	      92	  13332504 ns/op	  42.54 MB/s	 3631141 B/op	      51 allocs/op
BenchmarkDecodeRand-12      	     950	   1178515 ns/op	  13.90 MB/s	 3644460 B/op	      51 allocs/op
BenchmarkDecodeRand-12      	    1022	   1180043 ns/op	  13.88 MB/s	 3644463 B/op	      51 allocs/op
BenchmarkDecodeRand-12      	    1014	   1174522 ns/op	  13.95 MB/s	 3644462 B/op	      51 allocs/op
BenchmarkWiktionary-12      	       1	160628872250 ns/op	  65.36 MB/s	367217144 B/op	  542483 allocs/op
BenchmarkWiktionary-12      	       1	165487979792 ns/op	  63.44 MB/s	367217160 B/op	  542483 allocs/op
BenchmarkWiktionary-12      	       1	163573653500 ns/op	  64.19 MB/s	367217160 B/op	  542483 allocs/op
```